Code Cleanup and version bump. Betalgo.OpenAI 7.1.4 Betalgo.OpenAI.Utilities 7.0.2
