+++ OpenAI.Playground/TestHelpers/CompletionTestHelper.cs
