+++ OpenAI.SDK/Managers/OpenAIFile.cs
@@ -1,7 +1,6 @@
 ﻿using System.Net.Http.Json;
 using OpenAI.Extensions;
 using OpenAI.Interfaces;
-using OpenAI.ObjectModels.ResponseModels;
 using OpenAI.ObjectModels.ResponseModels.FileResponseModels;
 using OpenAI.ObjectModels.SharedModels;
