+++ OpenAI.SDK/Managers/OpenAIService.cs
@@ -8,13 +8,13 @@ namespace OpenAI.Managers;
 //TODO Find a way to show default request values in documentation
 public partial class OpenAIService : IOpenAIService, IDisposable
 {
+    private readonly bool _disposeHttpClient;
     private readonly IOpenAiEndpointProvider _endpointProvider;
     private readonly HttpClient _httpClient;
     private string? _defaultModelId;
-    private bool _disposeHttpClient;
 
     [ActivatorUtilitiesConstructor]
-    public OpenAIService(IOptions<OpenAiOptions> settings,HttpClient httpClient)
+    public OpenAIService(IOptions<OpenAiOptions> settings, HttpClient httpClient)
         : this(settings.Value, httpClient)
     {
     }
@@ -58,6 +58,15 @@ public partial class OpenAIService : IOpenAIService, IDisposable
         };
 
         _defaultModelId = settings.DefaultModelId;
+    }
+
+    /// <summary>
+    ///     Method to dispose the HttpContext if created internally.
+    /// </summary>
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
     }
 
 
@@ -107,15 +116,6 @@ public partial class OpenAIService : IOpenAIService, IDisposable
     public string? GetDefaultModelId()
     {
         return _defaultModelId;
-    }
-
-    /// <summary>
-    ///     Method to dispose the HttpContext if created internally.
-    /// </summary>
-    public void Dispose()
-    {
-        Dispose(true);
-        GC.SuppressFinalize(this);
     }
 
     protected virtual void Dispose(bool disposing)
