+++ OpenAI.SDK/ObjectModels/Models.cs
@@ -67,9 +67,11 @@ public static class Models
         CodeCushmanV1,
 
         CodeDavinciV2,
+
         [Obsolete("Use Gpt_3_5_Turbo instead")]
         ChatGpt3_5Turbo,
         Gpt_3_5_Turbo,
+
         [Obsolete("Use Gpt_3_5_Turbo_0301 instead")]
         ChatGpt3_5Turbo0301,
         Gpt_3_5_Turbo_0301,
@@ -133,15 +135,16 @@ public static class Models
     public static string Gpt_4_32k_0314 => "gpt-4-32k-0314";
 
     /// <summary>
-    ///     Snapshot of gpt-4 from June 13th 2023 with function calling data. Unlike gpt-4, this model will not receive updates, 
-    ///     and will be deprecated 3 months after a new version is released.	
+    ///     Snapshot of gpt-4 from June 13th 2023 with function calling data. Unlike gpt-4, this model will not receive
+    ///     updates,
+    ///     and will be deprecated 3 months after a new version is released.
     ///     8,192 tokens	Up to Sep 2021
     /// </summary>
     public static string Gpt_4_0613 => "gpt-4-0613";
 
     /// <summary>
-    ///     Snapshot of gpt-4-32 from June 13th 2023. Unlike gpt-4-32k, this model will not receive updates, 
-    ///     and will be deprecated 3 months after a new version is released.	
+    ///     Snapshot of gpt-4-32 from June 13th 2023. Unlike gpt-4-32k, this model will not receive updates,
+    ///     and will be deprecated 3 months after a new version is released.
     ///     32,768 tokens	Up to Sep 2021
     /// </summary>
     public static string Gpt_4_32k_0613 => "gpt-4-32k-0613";
@@ -199,6 +202,7 @@ public static class Models
     /// </summary>
     [Obsolete("Use Gpt_3_5_Turbo instead, this is just a naming change,the field will be removed in next versions")]
     public static string ChatGpt3_5Turbo => "gpt-3.5-turbo";
+
     /// <summary>
     ///     Most capable GPT-3.5 model and optimized for chat at 1/10th the cost of text-davinci-003. Will be updated with our
     ///     latest model iteration.
@@ -207,7 +211,7 @@ public static class Models
     public static string Gpt_3_5_Turbo => "gpt-3.5-turbo";
 
     /// <summary>
-    ///     Same capabilities as the standard gpt-3.5-turbo model but with 4 times the context.	
+    ///     Same capabilities as the standard gpt-3.5-turbo model but with 4 times the context.
     ///     16,384 tokens	Up to Sep 2021
     /// </summary>
     public static string Gpt_3_5_Turbo_16k => "gpt-3.5-turbo-16k";
@@ -219,6 +223,7 @@ public static class Models
     /// </summary>
     [Obsolete("Use Gpt_3_5_Turbo_0301 instead, this is just a naming change,the field will be removed in next versions")]
     public static string ChatGpt3_5Turbo0301 => "gpt-3.5-turbo-0301";
+
     /// <summary>
     ///     Snapshot of gpt-3.5-turbo from March 1st 2023. Unlike gpt-3.5-turbo, this model will not receive updates, and will
     ///     only be supported for a three month period ending on June 1st 2023.
@@ -227,15 +232,15 @@ public static class Models
     public static string Gpt_3_5_Turbo_0301 => "gpt-3.5-turbo-0301";
 
     /// <summary>
-    ///     Snapshot of gpt-3.5-turbo from June 13th 2023 with function calling data. Unlike gpt-3.5-turbo, 
-    ///     this model will not receive updates, and will be deprecated 3 months after a new version is released.	
+    ///     Snapshot of gpt-3.5-turbo from June 13th 2023 with function calling data. Unlike gpt-3.5-turbo,
+    ///     this model will not receive updates, and will be deprecated 3 months after a new version is released.
     ///     4,096 tokens	Up to Sep 2021
     /// </summary>
     public static string Gpt_3_5_Turbo_0613 => "gpt-3.5-turbo-0613";
 
     /// <summary>
-    ///     Snapshot of gpt-3.5-turbo from June 13th 2023 with function calling data. Unlike gpt-3.5-turbo, 
-    ///     this model will not receive updates, and will be deprecated 3 months after a new version is released.	
+    ///     Snapshot of gpt-3.5-turbo from June 13th 2023 with function calling data. Unlike gpt-3.5-turbo,
+    ///     this model will not receive updates, and will be deprecated 3 months after a new version is released.
     ///     4,096 tokens	Up to Sep 2021
     /// </summary>
     public static string Gpt_3_5_Turbo_16k_0613 => "gpt-3.5-turbo-16k-0613";
