+++ OpenAI.SDK/ObjectModels/RequestModels/CompletionCreateRequest.cs
