+++ OpenAI.SDK/ObjectModels/ResponseModels/BaseResponse.cs
@@ -47,37 +47,34 @@ public record BaseResponse
 
 public class Error
 {
-    [JsonPropertyName("code")]
-    public string? Code { get; set; }
+    [JsonPropertyName("code")] public string? Code { get; set; }
 
-    [JsonPropertyName("param")]
-    public string? Param { get; set; }
+    [JsonPropertyName("param")] public string? Param { get; set; }
 
-    [JsonPropertyName("type")]
-    public string? Type { get; set; }
-    [JsonIgnore]
-    public string? Message { get; private set; }
-    [JsonIgnore]
-    public List<string?> Messages { get; private set; }
+    [JsonPropertyName("type")] public string? Type { get; set; }
+
+    [JsonIgnore] public string? Message { get; private set; }
+
+    [JsonIgnore] public List<string?> Messages { get; private set; }
 
     [JsonPropertyName("message")]
     [JsonConverter(typeof(MessageConverter))]
-    public object MessageObject 
-    { 
+    public object MessageObject
+    {
         set
         {
             switch (value)
             {
                 case string s:
                     Message = s;
-                    Messages = new List<string?> { s };
+                    Messages = new List<string?> {s};
                     break;
-               case List<object> list when list.All(i => i is JsonElement):
+                case List<object> list when list.All(i => i is JsonElement):
                     Messages = list.Cast<JsonElement>().Select(e => e.GetString()).ToList();
                     Message = string.Join(Environment.NewLine, Messages);
                     break;
             }
-        } 
+        }
     }
 
     public class MessageConverter : JsonConverter<object>
@@ -88,10 +85,12 @@ public class Error
             {
                 return reader.GetString();
             }
-            else if (reader.TokenType == JsonTokenType.StartArray)
+
+            if (reader.TokenType == JsonTokenType.StartArray)
             {
                 return JsonSerializer.Deserialize<List<object>>(ref reader, options);
             }
+
             throw new JsonException();
         }
