--- OpenAI.Utilities/EmbedStaticValues.cs
@@ -1,10 +1,0 @@
-﻿namespace OpenAI.Utilities
-{
-    public static class EmbedStaticValues
-    {
-        public const string NTokens = "NToken";
-        public const string Distances = "Distance";
-        public const string Embeddings ="Embedding";
-        public const string Text = "Text";
-    }
-}
