+++ OpenAI.Utilities/Embedding/EmbedStaticValues.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenAI.Utilities.Embedding;
+
+public static class EmbedStaticValues
+{
+    public const string NTokens = "NToken";
+    public const string Distances = "Distance";
+    public const string Embeddings = "Embedding";
+    public const string Text = "Text";
+}
