+++ OpenAI.Utilities/OpenAI.Utilities.csproj
@@ -10,7 +10,7 @@
 		<PackageProjectUrl>https://openai.com/</PackageProjectUrl>
 		<PackageIcon>OpenAI-Betalgo.png</PackageIcon>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-		<Version>7.0.1</Version>
+		<Version>7.0.2</Version>
 		<Authors>Tolga Kayhan, Betalgo</Authors>
 		<Company>Betalgo Up Ltd.</Company>
 		<Product>Utility tools for Betalgo.OpenAI</Product>
@@ -41,13 +41,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Betalgo.OpenAI" Version="7.1.3" />
 		<PackageReference Include="CsvHelper" Version="30.0.1" />
 		<PackageReference Include="MathNet.Numerics" Version="5.0.0" />
 		<PackageReference Include="Microsoft.Data.Analysis" Version="0.20.1" />
 	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\OpenAI.SDK\OpenAI.csproj" />
-	</ItemGroup>
-
 </Project>
