--- OpenAI.Utilities/TextEmbeddingData.cs
@@ -1,9 +1,0 @@
-﻿namespace OpenAI.Utilities
-{
-    public class TextEmbeddingData
-    {
-        public string FileName { get; set; }
-        public string Text { get; set; }
-        public int NToken { get; set; }
-    }
-}
